requirements.txt
pyrogram==2.0.100
tgcrypto==1.2.5
uvicorn==0.24.0
fastapi[all]==0.104.1
python-dotenv==1.0.0
aiofiles==23.2.1
aiohttp==3.9.1
curl_cffi==0.6.2
techzdl==1.2.6
tqdm==4.66.1
dill==0.3.7
python-multipart==0.0.6
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
aiohttp-cors==0.7.0